Removed old tests and console, added test for faucet from test DApp
