+++ test/appium/tests/marks.py
@@ -16,5 +16,4 @@ transaction = pytest.mark.transaction
 wallet = pytest.mark.wallet
 sign_in = pytest.mark.sign_in
 skip = pytest.mark.skip
-console = pytest.mark.console
 logcat = pytest.mark.logcat
